polynomal/config: extract api and application config into own classes
In an effort to normalize related config properties, I created two new `Config::` subclasses, `Config::Api` & `Config::Application`.

`Config::Api` keeps state of the external api config. For example, the state of consists of the API host, port, and api key for polynomal.com.

`Config::Application` keeps state of the "parent" application or process.  State consists of the application name and environment. For Ruby, the state is set programmatically through  `Polynomal.configure`. For Rails, the state can be set automatically or programmatically.

+ Update specs for `config.rb`
+ Update `lib/polynomal/transporter/http.rb` to use new config classes